requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.6
APScheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0
sqlite3
Werkzeug==2.3.7
asyncio
aiofiles==23.2.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3